+++ src/components/content/conditions/helpers/history.tsx
@@ -7,7 +7,7 @@ import { ConditionModel } from "@/fhir/models";
 import { capitalize, startCase } from "@/utils/nodash";
 import { QUERY_KEY_CONDITION_HISTORY } from "@/utils/query-keys";
 
-export function useConditionHistory(condition: ConditionModel) {
+export function useConditionHistory(enableFQS: boolean, condition: ConditionModel) {
   return useHistory({
     resourceType: "Condition",
     model: condition,
@@ -15,6 +15,7 @@ export function useConditionHistory(condition: ConditionModel) {
     valuesToDedupeOn,
     getSearchParams,
     getHistoryEntry,
+    enableFQS,
   });
 }