+++ src/components/content/document/patient-documents.tsx
@@ -20,7 +20,7 @@ function PatientDocumentsComponent({ className, enableFQS = false }: PatientDocu
   const containerRef = useRef<HTMLDivElement>(null);
   const { featureFlags } = useCTW();
 
-  const patientDocumentQuery = usePatientDocument(enableFQS);
+  const patientDocumentQuery = usePatientDocument();
   const { viewOptions, defaultView } = getDateRangeView<DocumentModel>("dateCreated");
   const { data, setViewOption } = useFilteredSortedData({
     defaultView,
@@ -30,6 +30,7 @@ function PatientDocumentsComponent({ className, enableFQS = false }: PatientDocu
   const openDetails = useResourceDetailsDrawer({
     header: (m) => `${m.dateCreated} - ${m.title}`,
     details: documentData,
+    enableFQS,
   });
 
   return (