+++ src/components/content/story-helpers/mocks/requests/requests.ts
@@ -1,5 +1,6 @@
 import { rest } from "msw";
 import { binary } from "../resources/binary";
+import { unleashVariant } from "../resources/unleash";
 
 export const mockBinaryGet = () => [
   rest.get("https://api.dev.zusapi.com/fhir/Binary/:BinaryId", async (_, res, ctx) =>
@@ -11,3 +12,11 @@ export const mockBinaryGet = () => [
     res(ctx.status(500))
   ),
 ];
+
+export const mockUnleashFQSEnabledGet = (resourceType: string) =>
+  rest.get("https://unleash-proxy-prod.zusapi.com/*", async (_, res, ctx) =>
+    res(
+      ctx.status(200),
+      ctx.json(unleashVariant("test", "conditions", { [resourceType]: true }, true))
+    )
+  );