+++ src/components/core/providers/ctw-provider.tsx
@@ -5,10 +5,12 @@ import {
   AuthenticationProviderProps,
 } from "./authentication/authentication-provider";
 import { CTWStateContext, FeatureFlags } from "./ctw-context";
+import { FeatureFlagProvider } from "./feature-flag-provider";
 import { TelemetryProvider } from "./telemetry/telemetry-provider";
 import { ThemeProvider, ThemeProviderProps } from "./theme/theme-provider";
 import { Env } from "./types";
 import { version } from "../../../../package.json";
+import { OnResourceSaveCallback } from "@/fhir/action-helper";
 import { queryClient } from "@/utils/request";
 
 type CTWProviderProps = {
@@ -17,6 +19,7 @@ type CTWProviderProps = {
   enableTelemetry?: boolean;
   ehr?: string;
   featureFlags?: FeatureFlags;
+  onResourceSave?: OnResourceSaveCallback;
 } & ThemeProviderProps &
   AuthenticationProviderProps;
 
@@ -47,6 +50,7 @@ export function CTWProvider({
   headers,
   authToken,
   authTokenURL,
+  onResourceSave,
 }: PropsWithChildren<CTWProviderProps>) {
   useEffect(() => {
     window.CTWComponentLibrary = {
@@ -59,23 +63,26 @@ export function CTWProvider({
       env,
       builderId,
       featureFlags,
+      onResourceSave,
     }),
-    [env, builderId, featureFlags]
+    [env, builderId, featureFlags, onResourceSave]
   );
 
   return (
     <ThemeProvider theme={theme} locals={locals}>
       <AuthenticationProvider headers={headers} authToken={authToken} authTokenURL={authTokenURL}>
-        <TelemetryProvider
-          env={env}
-          builderId={builderId}
-          ehr={ehr}
-          enableTelemetry={enableTelemetry}
-        >
-          <CTWStateContext.Provider value={providerState}>
-            <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
-          </CTWStateContext.Provider>
-        </TelemetryProvider>
+        <FeatureFlagProvider>
+          <TelemetryProvider
+            env={env}
+            builderId={builderId}
+            ehr={ehr}
+            enableTelemetry={enableTelemetry}
+          >
+            <CTWStateContext.Provider value={providerState}>
+              <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+            </CTWStateContext.Provider>
+          </TelemetryProvider>
+        </FeatureFlagProvider>
       </AuthenticationProvider>
     </ThemeProvider>
   );