+++ src/components/core/providers/use-ctw.ts
@@ -28,6 +28,11 @@ export function useCTW() {
       builderId: context.builderId ?? claimsBuilderId(authToken) ?? "",
       contextBuilderId: context.builderId,
       fetchFromFqs: getFetchFromFqs(context.env, authToken, context.builderId),
+      onResourceSave: (resource, action, error) => {
+        if (context.onResourceSave) {
+          context.onResourceSave(resource, action, error);
+        }
+      },
     };
     return requestContext;
   }, [context, getAuthToken]);